Fix the save method.

+++ testing/test_model.py
@@ -44,9 +44,11 @@ class test_model2(unittest.TestCase):
     def setUp(self):
         tf.set_random_seed(0)
         self.m = SquareModel2()
-        self.filename = 'saved_file'
-        if os.path.exists(self.filename):
-            os.remove(self.filename)
+        self.filename = './saved_file.dat'
+        self.files = [self.filename+'.data-00000-of-00001',
+        self.filename+'.index', self.filename+'.meta', 'checkpoint']
+        if os.path.exists(self.filename): os.remove(self.filename)
+        [os.remove(f) for f in self.files if os.path.exists(f)]
 
     def test_initialize_finalize(self):
         optimizer = self.m.likelihood()
@@ -75,7 +77,7 @@ class test_model2(unittest.TestCase):
         self.m.q = np.ones((2,3))
         self.m.save(self.filename)
         # make sure the file is generated
-        self.assertTrue(os.path.exists(self.filename))
+        self.assertTrue(os.path.exists(self.files[0]))
         # make sure the model.restore()
         self.m2 = SquareModel2()
         self.m2.restore(self.filename)
@@ -84,21 +86,15 @@ class test_model2(unittest.TestCase):
         # make sure initialize() does not change the value
         self.m2.initialize()
         self.assertTrue(np.allclose(self.m2.q.value, np.ones((2,3))))
-        # remove the generated files
-        if os.path.exists(self.filename):
-            os.remove(self.filename)
-        if os.path.exists(self.filename+str('.meta')):
-            os.remove(self.filename+str('.meta'))
-        if os.path.exists('checkpoint'):
-            os.remove('checkpoint')
+        self.setUp()
 
     def test_parameterized(self):
         """ make sure save method for parameterized works fine """
         self.m.v.q_mu = np.ones((2*3))
         self.m.v.save(self.filename)
         # make sure the file is generated
-        self.assertTrue(os.path.exists(self.filename))
-        # make sure the model.restore()
+        self.assertTrue(os.path.exists(self.files[0]))
+        # make sure the model.restore() works
         self.m2 = SquareModel2()
         self.m2.v.restore(self.filename)
         self.assertTrue(np.allclose(self.m2.v.q_mu.value, np.ones((2*3))))
@@ -106,14 +102,7 @@ class test_model2(unittest.TestCase):
         # make sure initialize() does not change the value
         self.m2.initialize()
         self.assertTrue(np.allclose(self.m2.v.q_mu.value, np.ones((2*3))))
-        # remove the generated files
-        if os.path.exists(self.filename):
-            os.remove(self.filename)
-        if os.path.exists(self.filename+str('.meta')):
-            os.remove(self.filename+str('.meta'))
-        if os.path.exists('checkpoint'):
-            os.remove('checkpoint')
-
+        self.setUp()
 
 class MinibatchModel(hb.model.Model):
     def setUp(self):
